+++ CHANGELOG.md
@@ -6,7 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.2] - 2019-12-13
+
+### Changed
+- Remove middlewares in favour of aspects
+
 ### Fixed
+- Markup is not lost anymore when non-cached plugin on page (#27)
 - Don't show special doktypes in BreadcrumbList (#28)
 
 ## [1.4.1] - 2019-12-01
@@ -132,7 +138,8 @@ Initial release
 - View helpers for usage in Fluid templates
 
 
-[Unreleased]: https://github.com/brotkrueml/schema/compare/v1.4.1...HEAD
+[Unreleased]: https://github.com/brotkrueml/schema/compare/v1.4.2...HEAD
+[1.4.2]: https://github.com/brotkrueml/schema/compare/v1.4.1...v1.4.2
 [1.4.1]: https://github.com/brotkrueml/schema/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/brotkrueml/schema/compare/v1.3.1...v1.4.0
 [1.3.1]: https://github.com/brotkrueml/schema/compare/v1.3.0...v1.3.1